Globally allow Aqara devices to handle swapped ZCL header direction
