+++ zhaquirks/xiaomi/aqara/magnet_agl02.py
@@ -1,8 +1,6 @@
 """Quirk for Aqara T1 door sensor lumi.magnet.agl02."""
 
 from zigpy.profiles import zha
-import zigpy.types as t
-from zigpy.zcl import Cluster, foundation
 from zigpy.zcl.clusters.general import Basic, Identify, OnOff, Ota, PowerConfiguration
 from zigpy.zcl.clusters.security import IasZone
 
@@ -30,20 +28,6 @@ class MagnetT1(XiaomiCustomDevice):
         """Init."""
         self.battery_size = BatterySize.CR1632
         super().__init__(*args, **kwargs)
-
-    def _find_zcl_cluster(
-        self, hdr: foundation.ZCLHeader, packet: t.ZigbeePacket
-    ) -> Cluster:
-        """Find a cluster for the packet."""
-        assert packet.src_ep is not None
-        endpoint = self.endpoints[packet.src_ep]
-
-        # The `direction` field is incorrectly set, we should always route the packet
-        # to client `OnOff` cluster
-        if packet.cluster_id == OnOff.cluster_id:
-            return endpoint.out_clusters[packet.cluster_id]
-
-        return super()._find_zcl_cluster(hdr, packet)
 
     signature = {
         MODELS_INFO: [("LUMI", "lumi.magnet.agl02")],
